chore: Split up expander into more files
