+++ include/expander.h
@@ -6,7 +6,7 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/28 23:25:46 by ldulling          #+#    #+#             */
-/*   Updated: 2024/01/03 01:41:59 by ldulling         ###   ########.fr       */
+/*   Updated: 2024/01/03 15:26:43 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,16 +22,28 @@
 /* bad_substitution.c */
 bool	bad_substitution(char *str);
 
+/* expand_special_param.c */
+bool	expand_exit_code(char **str, size_t *i, int exit_code);
+
+/* expand_variable.c */
+bool	expand_variable(char **str, size_t *i, size_t offset, t_list *env_list);
+size_t	count_var_len(char *str);
+char	*get_replacement(char *var, t_list *env_list);
+
 /* expander.c */
 bool	ft_expander(char **str, t_shell *shell);
-void	skip_to_dollar_not_in_single_quotes(char *str, size_t *i);
 
 /* expander_utils.c */
+size_t	count_offset(char *str);
+size_t	count_replace_len(char *str);
 void	free_and_reset(char *dup);
 bool	is_open_pair(unsigned char c, int operation);
+void	skip_to_dollar_not_in_single_quotes(char *str, size_t *i);
 
 /* parameter_expansion.c */
 bool	parameter_expansion(char **str, t_shell *shell);
+bool	expand_parameter(char **str, size_t *i, t_shell *shell);
+bool	handle_dollar_quotes(char **str, size_t *i);
 
 /* quote_removal.c */
 bool	quote_removal(char **str);
