+++ source/frontend/expander/expander_utils.c
@@ -6,11 +6,48 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 01:01:07 by ldulling          #+#    #+#             */
-/*   Updated: 2024/01/03 01:42:15 by ldulling         ###   ########.fr       */
+/*   Updated: 2024/01/03 15:23:14 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "expander.h"
+#include "utils.h"
+
+size_t	count_offset(char *str)
+{
+	size_t	offset;
+
+	offset = 0;
+	if (*str == '$')
+	{
+		offset++;
+		str++;
+	}
+	if (*str == OPENING_BRACE)
+		offset++;
+	return (offset);
+}
+
+size_t	count_replace_len(char *str)
+{
+	size_t	replace_len;
+
+	replace_len = 0;
+	if (str[replace_len] == '$')
+		replace_len++;
+	if (str[replace_len] == OPENING_BRACE)
+	{
+		while (str[++replace_len] != CLOSING_BRACE)
+			;
+		replace_len++;
+	}
+	else if (str[replace_len] == '?')
+		replace_len++;
+	else
+		while (is_valid_varname(str[replace_len]))
+			replace_len++;
+	return (replace_len);
+}
 
 void	free_and_reset(char *dup)
 {
@@ -27,4 +64,19 @@ bool	is_open_pair(unsigned char c, int operation)
 	else if (operation == RESET)
 		is_open_pair[c] = false;
 	return (is_open_pair[c]);
+}
+
+void	skip_to_dollar_not_in_single_quotes(char *str, size_t *i)
+{
+	while (str[*i])
+	{
+		if (str[*i] == '$')
+			return ;
+		else if (str[*i] == '"')
+			is_open_pair('"', TOGGLE);
+		else if (str[*i] == '\'' && !is_open_pair('"', GET))
+			if (!skip_to_same_quote(str, i))
+				return ;
+		(*i)++;
+	}
 }
